Minor Additions
Small adds to begin to create a plug class. Need to decide if there is a
single plug object with a BiMap or just every plug create is its own
object. BiMap with a single plug option may be easier to work with as it
is not a progression for chars but rather a change and if the char isn't
in the BiMap it is returned as normal to transit the wheels